backend/voice/requirements.txt
openai>=1.52.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
numpy
